requirements.txt
asgiref==3.8.1
attrs==24.2.0
click==8.1.7
Django==5.1.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
et_xmlfile==2.0.0
gunicorn==23.0.0
h11==0.14.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
openpyxl==3.1.5
packaging==24.2
psycopg2-binary==2.9.10
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
sqlparse==0.5.2
uritemplate==4.1.1
